Prevent blowing up the callstack
This commit transforms the code such that, in the event of receiving a
tick far ahead of time from the last tick, the callstack does not blow
up with a `RangeError`.

One problem was located in `applyTime_`, which was recursively calling
itself.  While this is supposed to work without any problem, that call
was not tail-call optimized by the compiler and was essentially a time
bomb.

The solution that was applied to correct this issue was to extract the
simulations (both computation and storage) to the top-level `applyTime`
function, and (left-)fold over that list to apply transformations
successively.

One quick win was to encapsulates each single simulation (a tuple
composed of the pasteis simulation and the dought cost simultion) in a
single type.

    +type alias Simulations =
    +    { pasteisSimulation : Float
    +    , doughCostSimulation : Float
    +    }

So building a list of Simulations data structures was essentially a
matter of applying `map2` with that Simulation constructor (function
that has a signature : Float -> Float -> Simulations).