Use an extensible record to enable row polymorphism
It was possible to run both `runPasteis` and `runMegaPasteis`
interchangeably on `PasteisModule` and `MegaPasteisModule` as they were
defined exactly the same way (same attributes).

We used to have two different functions but for the wrong reason: the
type system was totally cool with this and we could have used only one
function, but it would have been an error as soon as one of these two
structures start diverging.

One solution to this was to leverage row polymorphism with the use of an
extensible record, which makes it possible to define a type that only
requires a subset of the information defined for any "production module"
(something that makes stuff, in this app).