Isolate Business entirely
This commit isolates the Business module entirely by hidding all the
logic in its Internals module.  This module should never be used outside
of any Business-related modules (View, Update, etc).

We're using the `Business.Update.update` function to send commands to
that module, and the module will return a list of top-level commands if
it finds it necessary.